pyproject.toml
[project]
name = "dictathesis"
version = "0.1.0"
description = "Smart voice dictation for thesis writing — one-handed, two-pass AI pipeline"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "httpx>=0.27.0",
    "pynput>=1.7.6",
    "pyperclip>=1.8.2",
    "pystray>=0.19.5",
    "Pillow>=10.0.0",
    "webrtcvad>=2.0.10",
]

[project.scripts]
dictathesis = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.15.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "PIE", # flake8-pie
    "SIM", # flake8-simplify
]
ignore = [
    "B008",   # function call in default argument
    "SIM105", # use contextlib.suppress (readability preference)
    "E501",   # line too long — long lines in prompt strings are intentional
]

[tool.ruff.lint.isort]
known-first-party = [
    "api_client", "audio", "hud", "injector", "pipeline",
    "prompt", "settings_store", "settings_ui", "tray",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
